+++ src/main/java/com/dart/api/application/review/ReviewService.java
@@ -6,16 +6,18 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 
-import lombok.RequiredArgsConstructor;
 import com.dart.api.domain.gallery.entity.Gallery;
 import com.dart.api.domain.gallery.repo.GalleryRepository;
 import com.dart.api.domain.review.entity.Review;
 import com.dart.api.domain.review.repo.ReviewRepository;
+import com.dart.api.dto.page.PageInfo;
+import com.dart.api.dto.page.PageResponse;
 import com.dart.api.dto.review.request.ReviewCreateDto;
-import com.dart.api.dto.review.response.PageInfo;
-import com.dart.api.dto.review.response.PageResponse;
+import com.dart.api.dto.review.response.ReviewReadDto;
 import com.dart.global.error.exception.NotFoundException;
 import com.dart.global.error.model.ErrorCode;
+
+import lombok.RequiredArgsConstructor;
 
 @Transactional
 @RestController
@@ -33,13 +35,13 @@ public class ReviewService {
 	}
 
 	@Transactional(readOnly = true)
-	public PageResponse readAll(Long galleryId, int page, int size) {
+	public PageResponse<ReviewReadDto> readAll(Long galleryId, int page, int size) {
 		final Gallery gallery = galleryRepository.findById(galleryId)
 			.orElseThrow(() -> new NotFoundException(ErrorCode.FAIL_GALLERY_NOT_FOUND));
 		final Pageable pageable = PageRequest.of(page, size);
 		final Page<Review> reviews = reviewRepository.findAllByGalleryOrderByCreatedAtDesc(gallery, pageable);
 		final PageInfo pageInfo = new PageInfo(reviews.getNumber(), reviews.isLast());
 
-		return new PageResponse(reviews.map(Review::toReviewReadDto).toList(), pageInfo);
+		return new PageResponse<>(reviews.map(Review::toReviewReadDto).toList(), pageInfo);
 	}
 }
