+++ src/main/java/com/dart/api/presentation/ReviewController.java
@@ -9,11 +9,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.dart.api.application.review.ReviewService;
+import com.dart.api.dto.page.PageResponse;
+import com.dart.api.dto.review.request.ReviewCreateDto;
+import com.dart.api.dto.review.response.ReviewReadDto;
+
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import com.dart.api.application.review.ReviewService;
-import com.dart.api.dto.review.request.ReviewCreateDto;
-import com.dart.api.dto.review.response.PageResponse;
 
 @RestController
 @RequiredArgsConstructor
@@ -29,7 +31,7 @@ public class ReviewController {
 	}
 
 	@GetMapping("/{gallery-id}")
-	public PageResponse readAll(
+	public PageResponse<ReviewReadDto> readAll(
 		@PathVariable(name = "gallery-id") Long galleryId,
 		@RequestParam(defaultValue = "0") int page,
 		@RequestParam(defaultValue = "10") int size
